TCM_Backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Pillow==10.4.0